+++ middleman-core/lib/middleman-core/load_paths.rb
@@ -7,7 +7,7 @@ module Middleman
       @_is_setup ||= begin
 
         # Only look for config.rb if MM_ROOT isn't set
-        if !ENV['MM_ROOT'] && (found_path = findup('config.rb'))
+        if !ENV['MM_ROOT'] && (found_path = findup('middleman.rb') || findup('config.rb'))
           ENV['MM_ROOT'] = found_path
         end